Move distance element to common package
