+++ web/src/common/components/Distance/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Distance";
